# .docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Banco de Talentos API
  description: API completa para gestão de candidatos, vagas e processo de recrutamento
  version: 1.0.0
  contact:
    name: Equipe de Desenvolvimento
    email: dev@empresa.com
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de Desenvolvimento
  - url: https://api.talentos.empresa.com/v1
    description: Servidor de Produção
...
security:
  - bearerAuth: []
paths:
...
  /jobs/{id}:
    get:
      tags: [jobs]
      summary: Obter detalhes de uma vaga
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Detalhes da vaga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

    put:
      tags: [jobs]
      summary: Atualizar vaga
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdateRequest'
      responses:
        '200':
          description: Vaga atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
...
    JobListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        company:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        contractType:
          type: string
        workModel:
          type: string
        status:
          type: string
        candidateCount:
          type: integer
        expectedHireDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    Job:
      allOf:
        - $ref: '#/components/schemas/JobListItem'
        - type: object
          properties:
            description:
              type: string
            requirements:
              type: string
            workCity:
              type: string
            workState:
              type: string
            salaryMin:
              type: number
            salaryMax:
              type: number
            startDate:
              type: string
              format: date
            headhunter:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                user:
                  $ref: '#/components/schemas/User'
                level:
                  type: string
            requiredSkills:
              type: array
              items:
                type: object
                properties:
                  skill:
                    $ref: '#/components/schemas/Skill'
                  isRequired:
                    type: boolean
                  proficiencyLevel:
                    type: string
    JobCreateRequest:
      type: object
      required: [title, description, companyId, contractType, workModel]
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        companyId:
          type: string
          format: uuid
        headhunterId:
          type: string
          format: uuid
        contractType:
          type: string
        workModel:
          type: string
        workCity:
          type: string
        expectedHireDate:
          type: string
          format: date
        salaryMin:
          type: number
        salaryMax:
          type: number
    JobUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/JobCreateRequest'
...